Test for a recheck opts regression caused by types-first
Summary:
Under types-first, the merge dependency graph gets calculated (at least partially) based on the implementation dependency graph, but we actually use the sig dependency graph to topologically sort components and schedule work. All components which are not marked as blocked by another component are immediately merged, so there is no opportunity to apply recheck opts to them.

This leads to the strange situation illustrated in this diff, where **adding** a sig dependency edge means that Flow does less work in a recheck than it otherwise would.

Note: I turned on abstract locations here because it's an easy way to get recheck opts to fire, but this is not specific to abstract locations.

Reviewed By: gabelevi

Differential Revision: D17636577

fbshipit-source-id: 8f0f6f5820c8485471529cb40b1f4ba3156f1b3e