Fix source of flakiness for lsp/completion test
Summary:
I noticed that {D17584973} failed this test. I think it was a source of flakiness. Most of the test cases look like

1. Add `jsx.js` (which has syntax errors)
2. Start the lsp server and server
3. Send a completion request
4. Assert the response matches.

When the lsp starts up, it will see the syntax errors and send the `textDocument/publishDiagnostics` message to the client. Usually this happens between steps 2 and 3. But if it happens after 3 then the step 4 assertion would fail like this:

```

+ textDocument/publishDiagnostics,textDocument/completion{"isIncomplete":false,"items":[{"label":"a","kind":6,"detail":"number","inlineDetail":"number","insertTextFormat":1}]}
- textDocument/completion{"isIncomplete":false,"items":[{"label":"a","kind":6,"detail":"number","inlineDetail":"number","insertTextFormat":1}]}
```

(Full logs: P113545680)

Easy fix. Just ignore those messages

Reviewed By: nmote

Differential Revision: D17602294

fbshipit-source-id: 7c6f2a7ae84f5d3caffae90fdaacc0e57ea46775