[tool] upgrade to jest 23
Summary:
we're using a very old version of jest. i'm trying to get us onto jest 24, but that also requires updating to babel 7, so updating to jest 23 as a stepping stone.

I looked at the [BREAKING] changes in https://github.com/facebook/jest/blob/master/CHANGELOG.md and none of them stand out to me as problematic. and the tests still pass, so...

Reviewed By: gabelevi

Differential Revision: D17447565

fbshipit-source-id: d6579d713d94f5a5298a94dc50caa30efcc8c0e6