`persistent_handling_result` no longer is a `result` type, now always returns a response
Summary:
When a persistent command handler errored, we had two options:

1. If we had a request ID (sent from the IDE with each request), we would send an `ResponseMessage(id, ErrorResult(...))`
2. If we did not have a request ID (maybe a didOpen errored or something like that), we would send a `NotificationMessage(TelemetryNotification(...))`

So when I add a custom request/response, what should we do when it errors or is cancelled? Neither of those options are appropriate. So I plan to add a custom error response.

So I basically change the type that each handler returns from

`(response, error_info) result`

to

`response`.

Even on an error, the persistent command handler is responsible for building its own response.

Reviewed By: mroch

Differential Revision: D17490136

fbshipit-source-id: db0a91fad2680271ab476d41791231abbaba7b4f