A more general approach to uncaught exceptions in persistent command handlers
Summary:
`commandHandler.ml` would catch any uncaught exceptions and then return an LspResponse with the error. This is great, but doesn't generalize to uncaught exceptions in non-lsp requests.

Instead, let's send the fact that there was an uncaught exception. Then `flowLsp.ml` can look at the exception & the request and decide what to do.

Reviewed By: mroch

Differential Revision: D17506139

fbshipit-source-id: 7b261a80b859bd70b4a9037daa7103bedc17a23f