set_live_non_parse_errors_and_send should take only 1 file's errors
Summary:
Originally, my plan was to pass in a map of uris to errors. But really, our intention is to only set the errors for a single file. That is

1. `foo.js` changes
2. We run `check-contents` on `foo.js`
3. We set the live errors for `foo.js`

Technically `check-contents` can send us errors the blame to other files. But really we just want to set errors for `foo.js`. If we set errors for other files, they might not get cleared correctly when files are closed.

Reviewed By: mroch

Differential Revision: D17582938

fbshipit-source-id: 8e106a7bfac8c6888792375e5305c68e96646661