./tool windtunnel-prof should log more memory stats to ODS
Summary:
`flow check --profile` reports a lot of memory stats. Let's log them to ODS when we run `./tool windtunnel-prof` so that we can watch for wins and regressions.

In particular, I want to watch `flow.memory.TOTAL.worker_rss_avg.start`, which is the average size of a worker right before it performs a job. It seems like it is easy to regress by adding significant work to the OCaml toplevel. [Some more context here](https://fb.workplace.com/groups/floweng/permalink/2863236100391653/).

Reviewed By: mroch

Differential Revision: D17723572

fbshipit-source-id: 82b3fce0b797c16c893f7c55b5d97fe6ebf08aba