Delete module.resolver support
Summary:
`module.resolver` is a `.flowconfig` option that lets you shell out to an external script in order to resolve a module reference to a file. A malicious actor could replace that external script with something bad, and then Flow would execute that bad code. This is a security hole and this feature is getting deleted.

We experimented with this feature but it's dead internally (and hopefully externally).

To generate this diff I deleted the option from `.flowconfig` and then cleaned up all the build errors and warnings. I also grepped for references to `module.resolver` and `resolver`. Hopefully I got all the relevant code.

Reviewed By: mroch

Differential Revision: D17738452

fbshipit-source-id: 1376759b238fa360a97ef2f791caa1ee8c312c40