Fix fd leak in socket acceptor
Summary:
I regressed this in {D16841302}. When the `flow lsp` process tries connecting to the monitor process, we decline the connection if the server is still initializing. Before the regression, we'd spew some logs but at least close the fd. After the regression, we stopped spewing but we forgot to close the fd.

This means we were leaking ~1 fd per second during initialization. If the server then restarts a few times (maybe due to libdefs being edited), then we can leak 500 fds and users will see the "Too many clients" error and restarting will be the only fix.

Reviewed By: nmote

Differential Revision: D17750058

fbshipit-source-id: 611eb01b00ec1996dca93f0b12ad8d6daa9e43d5