[parser] don't consume token on class body error
Summary: when we expect a class body, it seems unlikely that someone would type something other than a `{`, intending it to be a `{` where the actual body follows that unexpected token. so, we'd recover better if we did not treat the unexpected token as a typo'd start of the class body.

Reviewed By: gabelevi

Differential Revision: D17113972

fbshipit-source-id: a4e41183068c5f8cc0b76b25a41f4d7d923aec1b