[hack] Correctly query for XHP classes using preceding colon
Summary:
While working on the new serverless IDE autocomplete tests, we discovered that XHP classes were not being correctly autocompleted.

The confusion was generally around stripping preceding namespace tokens.  In general, for hack classes with namespaces, we needed to remove the preceding backslash.  For XHP classes, we needed to retain the preceding colon.

In cases where the purpose of stripping namespace is unclear, I have added comments to create expectations.

In cases where the user is typing an XHP tag, the preceding colon is assumed.  This means that, in the case where the context identifies that a tag is being typed, we need to insert the preceding colon when querying our symbol index, but we need to remove the colon when producing results.

In order to keep some tests returning correct results, I needed to port the XHP determination logic from the sharedmem/trie "compute_complete_global" pathway to the parent "find_global_results" method, so that code could be shared between sharedmem/trie and sqlite.

Reviewed By: 2BitSalute

Differential Revision: D17744925

fbshipit-source-id: 72b3e0f0c34cb3fcabaea0a3baadb979da9af327