Call `do_live_diagnostics` when the lsp connects to the server
Summary:
When the LSP is already connected to the server, this is what happens when a didOpen comes in

1. LSP gets a didOpen
2. LSP parses the file and sets the live parse errors
3. LSP sends the didOpen to the server and requests live errors
4. Server returns the live errors
5. LSP sets the live non-parse errors

However, if the LSP is not connected to a server when a didOpen comes in, it doesn't do steps 3-5. Eventually, if the LSP connects to a server, it will create synthetic didOpen notifications and send those to the server.

After this change (and the later change which adds live non-parse errors), the LSP will also request the live errors when it connects.

Reviewed By: avikchaudhuri

Differential Revision: D17621748

fbshipit-source-id: fa73740cd9878d94f257428de98aaa81dc51f7e3