.flowconfig option to disable live non-parse errors
Summary:
In the future, we might find a bug that looks like this:

* `flow check` reports N errors & warnings for `foo.js`
* `flow check-contents` reports M errors & warnings for `foo.js`
* N != M

So maybe `check-contents` reports an error that `check` does not, or vice versa. This means as you make changes to a file, you might miss errors or see an error that isn't really there. That would be bad.

So let's make it possible to turn off this behavior temporarily until we can fix it.

Reviewed By: jbrown215

Differential Revision: D17668802

fbshipit-source-id: 91d0880909f1f0ce8364327a8d39513ebd3fe73e