Separate to_check and to_merge sets
Summary:
We've been using the same set of files to drive both the merge and the check phase of types-first. Unfortunately, this prevents us from implementing optimizations specific to one phase.

In particular, we only need to merge files which are transitive *signature* dependencies of the changed files.

This diff does not perform any optimizations, it simply introduces the concept that `to_check` and `to_merge` are distinct.

Reviewed By: gabelevi

Differential Revision: D17739253

fbshipit-source-id: 2461e162290ef479bd6e12bf6b67e28280f2a42f