Unit tests for calc_all_dependents
Summary:
This function behaves in a way that makes sense, but isn't necessarily obvious from the name.

Basically, instead of actually calculating all the dependents (using either the sig dependency graph or the implementation dependency graph), it computes all of the transitive sig dependents, then adds all of the direct implementation dependents of that set. This gives the set of all files that need to be rechecked, since we need to check any file whose implementation depends on a signature that might have changed.

Reviewed By: gabelevi

Differential Revision: D17750398

fbshipit-source-id: fb47a16f7bca6bae7dd5c91d132956d6e5dacdc0