fix logging for requests when server disconnected
Summary:
when the LSP server receives a request from the IDE when it is disconnected from the Flow server, it responds with a `RequestCancelled` error and logs it as a `PERSISTENT_EXPECTED_ERROR`.

`PERSISTENT_EXPECTED_ERROR` is for "A non-command-related expected error, e.g. server shuts down". We use `PERSISTENT_COMMAND_SUCCESS` with an `"Error"` key to respond to other requests with "expected" errors.

Since it's a request from the LSP client, the latter is what we want. Also, `PERSISTENT_COMMAND_SUCCESS` logs more stuff, like `persistent_context` and timing information.

For non-requests like notifications, we still fall back to `PERSISTENT_EXPECTED_ERROR`.

Reviewed By: gabelevi

Differential Revision: D17866333

fbshipit-source-id: 09022ed9a8fd4dc037ea832054321d90b728a33a