[types-first] Skip implementation of declared functions
Summary:
In types-first the signature generator gathers a list of statements that correspond
to each exported identifier. When an identifier is declared multiple times before being
exported, the signature generator includes *all* these declarations in the signature.
This differs from the behavior of classic with similar exports. There, Flow typically
only considers the first declaration and ignores the rest. A notable exception is the
case of `declare function`, where all declarations are included, but not a potentially
trailing implementation.

This diff bridges the above difference. When computing the generated signature in
types-first, we:
* fix the kind of the export (e.g. func. declaration, class, etc.) on the first occurrence.
* include the first declaration (based on location) and ignore all other declarations,
unless we have multiple function declarations, which are all included.

For example for this code:
```
var Foo: number = 1;
class Foo {};
export {Foo};
```
`Foo` is exported as a number in both modes (types-first used to export this as empty/any).

Reviewed By: samwgoldman

Differential Revision: D17874223

fbshipit-source-id: 3027e9a73ec5ba0dfec834470f5792a562a6ae4e