Fix some of the lsp/diagnostics tests
A couple of the wait conditions were wrong, causing the test runner to
wait the full timeout. The expectation was correct, so the test would
pass, but it would take too long.

I bumped all of the wait conditions to 60 seconds, but one of them was
expected to fail. In my opinion, this is a bad test, because it needs to
be slow by design. I just made the timeout smaller.

I deleted the last test, because it had no assertion and was actually
failing. I could have made given this test an assertion, but concluded
that it wasn't worth it. This should be a unit test.