[hack] factor out common rpc error types
Summary:
I have a philosophical belief that (1) we should use the Result monad for expected codepaths and use exceptions for code-paths that we the coder didn't anticipate, (2) for code interacting with the outside world over pipes or file-descriptors, we your code should anticipate and handle anything. Therefore, all such code should use the Result monad to indicate connection failures and the like.

I already wrote hh_mapreduce with that principal in mind for communication between orchestrator and prototype/worker. The next job is to apply the same principal for communication between worker and decl-service.

This diff moves the IPC result monad that I've been using into Marshal_tools, where it will be able to be used by both.

Reviewed By: 2BitSalute

Differential Revision: D17968596

fbshipit-source-id: b84be3433a61a1983efd0d0c45e4fef188ecb08b