Base to_merge solely on the sig dependency graph
Summary:
"Merge" and "check" are the two phases of typechecking under types-first. The purpose of the merge phase is to create sig contexts for each file, which the check phase consumes when typechecking the body of each file. Signatures are computed by extracting only the pieces of a file which are relevant to its externally-visible interface.

Importantly, many imports are only used in the body of a file, and do not appear in its signature. This means that the signature dependency graph is sparser than the implementation dependency graph.

Previously, when running a recheck, we passed the same set of files to the merge phase and the check phase[1]. Because there was no separation, this means that we had to merge everything that we had to check.

When files change, consider what we have to do to find any downstream errors that the changes may have caused[2]:
* The merge phase must make sure that all sig contexts are up-to-date. In order to do that, it has to remerge any file whose **signature** depends (transitively) on a changed file, as well as the changed files themselves.
* The check phase must check all files which might have been affected by a changed signature. Because the set of merged files is by definition the set of files that might have a changed signature, we have to take the set we computed for merge and add to it every file whose body depends on a file in that set.

Today, we compute the set of files we need to check, and merge all of those files, too. This is wasteful, because for many of them we know that their signature could not have possibly changed but we remerge them anyway. Because of how recheck opts are implemented in merge, this causes additional weird behavior, documented in D17636577.

This diff cuts the set of files that we consider for merge down to the set described above[3]. Manual testing suggests that this cuts down recheck times significantly.

I have this behind a temporary flag to mitigate rollout risk as well as give us an easy way to measure the impact of the change (when combined with the rollouts feature).

[1] This is not entirely true -- in lazy mode, we have to merge dependencies of changed files and dependencies of dependents of changed files. This diff didn't change that logic (except to omit dependencies from the `to_check` set where previously they were dropped downstream), so I'm leaving dependencies out of the discussion for the sake of simplicity.

[2] Again for the sake of simplicity, I am leaving recheck opts out. I'm discussing the sets of files that we *consider* for merge and check, but we might not end up merging/checking all of them due to recheck opts.

[3] Once again, this isn't quite true. We include all implementation **direct** dependents as part of the merge and check sets. These files get added much earlier in the process than the logic I've modified in this diff, and it isn't at all obvious from looking at what I've changed here. Trimming that set is next on my list.

Reviewed By: gabelevi

Differential Revision: D17821995

fbshipit-source-id: 70f89600d154195e4b7c5a46b0359388304f4d14