[website] Get rid of pre-0.89 react docs
Summary:
These docs are outdated and use patterns we no longer encourage. Instead of having to keep these docs up to do date, I want to get rid of the entirely.

We're now almost 20 versions past 0.89, so these docs hold little relevance now.

The topics covered in both versions of the docs are the same, minus the redundant parts from the old types.

Reviewed By: panagosg7

Differential Revision: D17062661

fbshipit-source-id: 624f98c00c3104cbd5cabc0365ba4e42da179c97