Use JS-friendly offsets in parser output
Summary:
This makes the standalone Flow parser (the one published as the `flow-parser` package) output offsets in a way that is compatible with other JS tools. Details are in D18017775.

This does not affect the output of the `flow ast` command. I'm not sure what the best default is there. It might make sense to expose this as a flag.

Reviewed By: gkz

Differential Revision: D18017774

fbshipit-source-id: 42efd18ffee7c072bfda96d142561ffa85bacb55