[spreads] Add use_desc to optional_t to preserve reason desc in VoidT
Summary:
Optional properties generated by widening are funky because they are "synthetic" -- there is no syntax to point to that explains the optional property.

Currently, the reason used for the optional property is the property that is not missing in one of the branches. This is confusing, and causes error messages like (see the next diff for changes) :

```
... undefined [1] ...

References:

prop: 3
^^^^^^^ [1]
```

In order to make better error messages, we should instead point to the object that does _not_ contain the property and make the reason indicate that it may be missing. In order to do that, we need to make sure that the VoidT emitted by OptionalT flows carries that reason with it. Adding the `use_desc` flag here makes it possible.

Since I don't like tuples in Type.t, I changed OptionalT's payload into an inline record.

Reviewed By: avikchaudhuri

Differential Revision: D17932144

fbshipit-source-id: 5ede4442df48ab9937024f5ece995a66c1c4f473