[spreads] Move temporary objects to type spread
Summary:
Moves temporary objects onto the normal object type resolution mechanism, which brings temporary objects to the new spread semantics.

We have to do one hacky thing here: previously, the object type was create with `Obj_type.mk_with_proto` and always assumed `exact = true`. To make sure that invariant is preserved, I had to add a special case to the output of `convert` in the Temporary object case to manually set the flag to true.

This _should_ not be the case, but this type should also not exist to begin with and it just preserves the behavior that already exists.

I did not do the same for objects that go through spread-- but that doesn't matter much. Most of the new errors with Temporary objects are because they were given an object literal reason in the old semantics and they no longer do in the new semantics. It was unsound to give exported object types a literal reason, so this change absorbs part of the error diff involved in getting rid of temporary objects.

Reviewed By: avikchaudhuri

Differential Revision: D17101960

fbshipit-source-id: 402c6542dcfee6092236124dfb32fb1089e0ee11