[spreads] Use spread reason over slice reason in mk_object for spreads
Summary:
When only one object is spread, the reason in the slice may differs from the spread reason. If there are at least two objects being spread, `spread2` sets the resulting object's reason to the spread reason instead of the slice reason. There's no good reason to use the slice reason to create the object when there's only one spread operand-- you'll even notice that the error messages have improved. We should always use the spread reason!

In type spread, this change has no semantic difference. However, for values we use the object reason to determine if the object is a literal or not. When it is, we allow more relaxed subtyping rules, since it's unaliased. If we spread a single aliased object, we want to use the spread's reason, which is a new object literal reason, and not the original reason.

The added test fails before this change.

Reviewed By: avikchaudhuri

Differential Revision: D17698733

fbshipit-source-id: fa281ba64c7a5d1bcc4db45f36918ecc7a420b1d