[spreads] Add indirection to computed properties
Summary:
This diff changes the way we determine the type of an object with a computed property so that we can more easily move value spread onto the type spread mechanism. In particular, instead of directly calling SetElemT on the `result` of an object value type calculation when we encounter a computed property, we can SetElemT on an intermediate empty unsealed object and then Obj.assign that object into the `result`. I need the indirection here so that I can catch when multiple lower bounds flow through the same computed property. We _do not_ allow this in the new spread semantics. In the next diff, I'll actually enforce that each CreateObjWithComputedPropT only gets triggered once.

## Paging it back in
When we are determining the type of an object value, say `{foo: 3, bar: 3, ...baz, [x]: true, [y]: false}`, we:
1. Create an empty unsealed object for the result
2. Add foo to result
3. Add bar to result
4. Obj.Assign(baz, result)
5. result ~> SetElemT (x, true)
6. result ~> SetElemT (y, true)

If `bar` is a union of types, or `x` or `y` has multiple lower bounds, this will combine all of the combinations of possible objects into a single resulting object with all of the keys. This is unsound, and it also does not play well with the spread redesign.

The spread calculation in Flow is a binary operator over two object types. To get that calculation working for value spread, we need to change the way we determine a type for an object value. The process to determine a type for `{foo: 3, bar: 3, ...baz, [x]: true, [y]: false}}` should instead be (approximately): `{...{foo: 3, bar: 3}, ...baz, ...{[x]: true}, ...{[y]: false}}`

This setup allows us to easily model value type calculation using the spread mechanism.

Reviewed By: avikchaudhuri

Differential Revision: D16968206

fbshipit-source-id: 65bab526f52c8881c1fb30329f9da28f6a13859a