[spreads] Better reasons for synthetic optional properties
Summary:
See the description in the previous diff.

This diff adds the reason desc and improves the error messages. See object_widening.exp for concrete examples!

Note: the reason for the widest object is always the spread reason. To get even better error messages here, we could keep track of where each property in the widest object came from so that we can point to objects missing a certain property.

Reviewed By: avikchaudhuri

Differential Revision: D17918033

fbshipit-source-id: 2cc9c4b8e2823e4ace5adc4799f21b3e74a05b56