[spreads] Basic widening
Summary:
A barebones widening implementation.

This widening algorithm:

1. Guarantees that a single object going through it will never cause new work on it's second time going through (good for perf, prevents cycles from infinitely creating new object ids)
2. Fires the same final constraint no matter what order we see constraints
3. Keeps objects exact whenever possible (desirable to keep error diff low)

However, this algorithm is O(n^2). Even so, spreads built on top of this widening algorithm is perf neutral.

## At a high level

Given a current widest object type `O` and a new object type `P`, we produce a new widest object
type `O'` by:
1. Merging the prop maps. If a property exists in (`O` and not `P`) or (`P` and not `O`) then it's added as an optional property.
If a property exists in both, we check if the value type in `P` is a subset of the value type in `O`. If it isn't, we create a new tvar and flow both the value type of `O` and `P` to it. If it is, we just take the value type from `O`
2. If `O` and `P` both have indexers, unify the key types to each other and the value types to each other. If only `O` xor `P` has an indexer, omit the indexer.
3. Set exactness to the conjunction of `O.flags.exact` and `P.flags.exact`.
4. If the props map, indexer, or exactness changed, emit a new widest type to the spread with the results of those changes and save `O'` as the new widest type, otherwise keep `O` and do not emit any flow to the spread.

## is_subset

is_subset is a wrapper around `TypeUtil.reasonless_eq` that adds some cases for our union-like types. The more complete `is_subset` is the more we can skip creating new tvars in widening.
NOTE: We don't consider OpenTs in the subset checking. However, once an OpenT is in the widest type, we will always flow the new types we encounter for that key to that OpenT and report that the type is unchanged.

## Follow up work

This diff does a few things poorly:
1. It widens union annotations (handled in D17794654). These should be skipped so that we can support things like spreading disjoint unions. Any perf issues caused by spreading these unions will be handled by linearity checking (below in the stack)
2. It creates new optional properties with poor reasons. The optional types will often point to the concrete property and say `undefined` even when no `undefined` is there explicitly. To fix this, I am going to add a specific reason for these synthetic optional properties and make sure those are used in the error messages. (fixed in D17794654)
3. Creates a lot of OpenTs with edges between them. Eventually the flows can get very expensive to process. We can avoid this by merging tvars via unification (fixed in D17798281).

Reviewed By: avikchaudhuri

Differential Revision: D17794655

fbshipit-source-id: 906774d0de4bedcfe9e9f60378ef009767ee3487