Check-contents-on-change should respect ignores/includes/file-exts/etc
Summary:
Long story short, if `flow check` doesn't report an error for `foo.js` then check-contents-on-change shouldn't either.

The one exception here is that `flow check` will only report errors in `node_modules/foo/foo.js` if it is a dependency of non-node-modules code. We don't do dependency graph calculations, so if `node_modules/foo/foo.js` has `flow` and you open it, then we'll report errors in it.

Reviewed By: mroch

Differential Revision: D18085991

fbshipit-source-id: a75b52d44fad3f5a2b6af3250f30f6be69ebcdd1