un-functorize Signature_builder_deps.ExpectedAnnotationSort
Summary:
now that `ExpectedAnnotationSort` does not contain a `Loc.t`, it doesn't need to be part of the `Signature_builder_deps.Make` functor. removing it gets rid of a bunch of boilerplate to convert from the `With_Loc` version to the `With_ALoc` version.

I couldn't figure out how to make ocaml let me define it in signature_builder_deps.ml so I moved it to its own file. But there are a lot of signature-related files now so I'm starting to make a parser_utils/signature_builder/ directory that we can move the rest into later.

Reviewed By: nmote

Differential Revision: D18059839

fbshipit-source-id: 43dae15b9a1d2259e6cb4f7642b86220bd0ae792