fix type of Error_message.EDuplicateModuleProvider location
Summary:
`Error_message.EDuplicateModuleProvider` (rightly) used `File_key.t` for the two conflicting module providers, but that meant that we had to synthesize the loc of the error in a couple places.

One of those was `aloc_of_msg`, which I am trying to generalize to work on `Loc.t` as well, so we can't call `ALoc.of_loc` there.

The (small) compromise is to store locs in `EDuplicateModuleProvider`, which is slightly unfortunate since sources are optional. But it's also more consistent for all errors to have locs, so ¯\_(ツ)_/¯

Reviewed By: dsainati1

Differential Revision: D18062321

fbshipit-source-id: f6717233509928ad1772e351713734eb50bcfc09