split EStrictLookupFailed into EBuiltinLookupFailed
Summary:
`EStrictLookupFailed` has 3 reasons attached to it, and it's not clear which is which. so, first of all, I changed it to an inline record.

secondly, two of the reasons are used for missing prop errors: `reason_prop`, which refers to the member expression, and `reason_obj` which refers to the object whose prop is being looked up. the third reason is only used if it is a "builtin reason", in which case the message is special-cased to refer to global variables, and the other two reasons aren't used.

thirdly, `aloc_of_msg` had to call `is_builtin_reason` to figure out the location of the error, which requires inspecting `ALoc`s, which I want to avoid so that I can make `aloc_of_msg` generic on `'loc`.

so, I split it into two kinds of errors, adding `EBuiltinLookupFailed`, and branched the callsites based on that third reason.

NOTE: I'm pretty sure some of the callers will never get a builtin reason and don't need to handle that case, but I want to leave that for future work since it is not obvious how to prove that right now.

NOTE: I also didn't create a helper function because each is slightly different, and it would continue to conflate the two types of errors as much as it did before, with the 3 confusing reasons and a bunch of unused extra params.

Reviewed By: jbrown215

Differential Revision: D18062470

fbshipit-source-id: 3f58b57867bf17db60a81a525a7f5e219cfa12bc