[fix] include loc in EInvalidPrototype
Summary:
`EInvalidPrototype` only included the reason of the lower bound that reached an
`ObjTestProtoT`. This caused a crash when the type of lower bound was invalid
(ie failed the proto-test) and did not cary a location. In particular the crash (in the
included test) was due to a `locationless_reason` called by the dummy prototype type
constructor.

This change introduces a location in `EInvalidPrototype` that is populated by
`ObjTestProtoT`. This use is more likely to have a proper location in its reason.
For all existing tests the two locations used to coincide, hence the no error diff.

Fixes [#8154](https://github.com/facebook/flow/issues/8154)

Reviewed By: nmote

Differential Revision: D18130624

fbshipit-source-id: 50c5e290177e0f2a6fec222de1a62ef55d07ca1f