add coverage test for function calls that return any
Summary:
we currently treat a function call that returns `any` as uncovered. this adds a test for that behavior.

I contend that this is wrong in cases where the return value is ignored, like in a statement:

if `x: () => any`, then `x();` does not leak `any` into the environment so I think it should be covered.

however, `const y = x();` should make `y` uncovered (but still not `x()`).

but if you do `foo(x())`, then the first argument is uncovered so then `x()` is uncovered.

Reviewed By: panagosg7

Differential Revision: D18137649

fbshipit-source-id: 53219a5c80b5fa5845c53ef946a3e183ef5d6722