Use let-open for error messages in error_message and debug_js
Summary:
When we initially enabled ocamlformat, it turned all `let-open`s into the shorthand version. This is annoying in some places because of the closing paren `)` required at the end of the match that needs to be moved around every time you add a new case (and is especially annoying during rebases). Additionally, it added unnecessary indentation.

Use `let-open` instead, now that we have updated the setting to not force the shorthand.

Reviewed By: jbrown215

Differential Revision: D18166165

fbshipit-source-id: fa1f3c03d6af499aeedd193b63009d0e4ebaba76