Move core autocomplete logic off of the type inference hooks
Summary:
Switch to using the typed AST instead of the type inference hooks to service autocomplete, like we've done with type-at-pos and get-def. I'll delete `autocomplete_set_hooks`/`autocomplete_unset_hooks` in a separate diff, since `statement.ml` seems to rely on them for ignoring `AUTO332`.

This also makes a subtle change to how the typed AST is constructed for JSX expressions:
```
<Component ...>
 ^^^^^^^^^
```
Previously, the type stored on the name of the JSX element (if the name was just an identifier) was the type of the entire element rather than the type of the component. I changed it because to autocomplete JSX attributes we need the type of the component to be stored in the typed AST. It seems like the previous was a mistake in the first place since it's inconsistent with what we do when the name is a member expression, and also it makes more sense for the name to be typed with the component it's referring to rather than with the type of the entire element.

Reviewed By: gkz

Differential Revision: D17964470

fbshipit-source-id: e5eb5b72a4baf1eeb387c327dce0d6086b5745e4