Simplify re-export code, drop "skip_dupes" flag
Summary:
The `skip_dupes` flag is only true for re-exported names, arising from the ES
module syntax `export * from 'm'`.

In this syntax form, re-exported names which conflict with locally export names
are ignored. Supporting this rule was the original motivation for the
`skip_dupes` flag.

Later, exports were categorized (`ExportValue`, `ExportType`, `ReExport`) which
making it possible to check that exported things are either types or values as
appropriate.

It turns out that this categorization subsumes the `skip_dupes` flag, since the
flag is true if and only if the export is a `ReExport`.

Note that the similar syntax form `export {x} from 'm'` is _not_ categorized as
`ReExport`, but rather `ExportValue`.

Reviewed By: panagosg7

Differential Revision: D18186094

fbshipit-source-id: 96f783c6d9bd7941273ae343bf0a8f1154d7ba07