[normalizer] Normalize type params
Summary:
`this` is represented internally as a type param on classes that get substituted. When that happens, we end up hitting a type alias with `TypeParamKind`.

Ideally, the typed ast would stop this from happening in the first place, but until then we can add special handling for these cases.

Type parameter hit this case as well with types that look like they come from `generate_tests`. To avoid displaying the bounds in the autocomplete results, we can reconstruct the type param's name from its reason.

Reviewed By: panagosg7

Differential Revision: D18143883

fbshipit-source-id: e8a0a7b9d35afbda398fcca2f9f770ba96a8394c