[autocomplete] Allow setting separate insert text for LSP, use this for JSX attributes
Summary:
LSP allows you to separately specify the text to insert from the label that is displayed. By default, if you don't set this separately, the label is used as the insert text.

Here I add a way to specify the insert text separately from the label, keeping existing behaviour (just have the label) if it is not specified.

There is a `insertText` property in the [LSP specification](https://microsoft.github.io/language-server-protocol/specifications/specification-3-14/#textDocument_completion), however it says:

> ... it is recommended to use `textEdit` instead

so that is what I have done.

As the first use of this, I have modified the autocomplete of JSX attributes to also add in the `=` after the attribute name, part of my effort to improve React-related autocomplete: https://fb.quip.com/GNLTA6E2VENY

Reviewed By: nmote

Differential Revision: D18066364

fbshipit-source-id: de239768b2b53235ad459f43962b731921000a5b